Create mod socket_listener
