+++ client/Cargo.toml
@@ -28,6 +28,7 @@ serde = { version = "1.0", features = ["derive"] }
 serde_yaml = "0.8.23"
 reqwest = { version = "0.11.10", features = ["stream"] }
 interprocess = "1.1.1"
+ctrlc = "3.2.1"
 
 [build-dependencies]
 tonic-build = "0.6.2"
